fix: global variable to fix testing flag.
