+++ service/utils/__init__.py
@@ -1,7 +1,4 @@
-import json
-
-from flask import Flask, g, session
-from google.cloud.firestore import CollectionReference
+from flask import Flask
 from werkzeug.exceptions import HTTPException
 
 from utils.db import init_db
@@ -9,14 +6,22 @@ from utils.db import init_db
 db = None
 
 
+class Config:
+    is_testing: bool = False
+
+
+config_g = Config()
+
+
 def create_app(test_mode=False):
     _app = Flask(__name__)
     global db
     if test_mode:
         db = init_db('fbConfigs_test.json')
-        with _app.app_context():
-            g.is_testing = True
-            session['is_testing'] = True
+        config_g.is_testing = True
+        # with _app.app_context():
+        #     g.is_testing = True
+        #     session['is_testing'] = True
     else:
         db = init_db('fbConfigs.json')
 
@@ -31,43 +36,6 @@ def create_app(test_mode=False):
     _app.register_error_handler(HTTPException, handle_http_exception)
 
     return _app
-
-
-def delete_collection(coll_ref: CollectionReference, batch_size=50):
-    """
-    Delete all documents in a collection.
-    :param coll_ref:
-    :param batch_size:
-    :return:
-    """
-    docs = coll_ref.limit(batch_size).stream()
-    deleted = 0
-
-    for doc in docs:
-        print(f'Deleting doc {doc.id} => {doc.to_dict()}')
-        doc.reference.delete()
-        deleted = deleted + 1
-
-    if deleted >= batch_size:
-        return delete_collection(coll_ref, batch_size)
-
-
-def load_collection(coll_ref: CollectionReference,
-                    data_path: str, id_key=None):
-    """
-    Add data to a collection from a json data file.
-    :param coll_ref: reference object to a firebase collection.
-    :param data_path:
-    :param id_key:
-    :return:
-    """
-    raw = json.loads(data_path)
-    data = raw.get('data')
-    for record in data:
-        if id_key in record:
-            coll_ref.document(record['id_key']).set(record)
-        else:
-            coll_ref.add(record)
 
 
 def format_headers(token=""):
